Use same structure as other connectors + combine channel and instanceName
